Let user decide if the original file should be kept or not (assume not)

+++ ajaxuploader/backends/thumbnail.py
@@ -7,8 +7,10 @@ from ajaxuploader.backends.local import LocalUploadBackend
 
 class ThumbnailUploadBackend(LocalUploadBackend):
     DIMENSION = "100x100"
+    KEEP_ORIGINAL = False
 
     def upload_complete(self, request, filename):
         thumbnail = get_thumbnail(self._path, self.DIMENSION)
-        os.unlink(self._path)
+        if not self.KEEP_ORIGINAL:
+            os.unlink(self._path)
         return {"path": settings.MEDIA_URL + thumbnail.name}
